+++ BaseProject/Source/Engine/Input/Internal/InputContext.h
@@ -7,7 +7,11 @@ namespace Engine::Input
 	{
 		// マウス制御
 		bool isMouseLockToCenter = false;
-		int deltaX = 0;
-		int deltaY = 0;
+
+		// このフレームのマウス移動量(感度を掛けた後)。
+		// 元は WM_INPUT で届いたデバイスのカウントで、整数のまま扱うと
+		// 感度を下げたときに端数が捨てられて動きが粗くなるため float で持つ。
+		float deltaX = 0.0f;
+		float deltaY = 0.0f;
 	};
 }
